Re-add EBI_SRAM sample code
The EBI interface on the NUC980 EV/IoT board is not connected to physical
SRAM, so the sample code cannot be executed correctly, and therefore we
removed it. However, due to user requests for EBI reference code, we have
re-added this sample code to the BSP. Please note that this sample code
is only provided as an example for EBI programming.

Change-Id: Ia4e25a350c243a6bad79ec1b2f5369376913e156